create new session article
Keras is a minimalisitc Python Deep Learning library that works on top of TensorFlow or Theano. It helps to quickly prototype Deep Learning models, and works seamlessly with both CPUs and GPUs. In this talk you will learn about how to build a simple image classification model using Keras. This is going to be more hands-on demo with code.

Speaker Bio: Malaikannan Sankarasubbu is a self taught Deep Learning practitioner. He attributes most of his learning to MOOC, he started on this journey by taking Andrew Ng's Machine Learning Course in 2012. He has been a avid developer of DL models for a past few years now. He is founder and CTO of datalog.ai.